requirements.txt
Flask==1.1.2
bs4
requests
APScheduler
gitPython
tqdm
google-cloud-storage